feat: add global Snackbar provider
