+++ src/main.jsx
@@ -7,23 +7,27 @@ import { RoomProvider } from './RoomContext.jsx'
 import { ThemeProvider } from './ThemeContext.jsx'
 import { WeatherProvider } from './WeatherContext.jsx'
 import { UserProvider } from './UserContext.jsx'
+import SnackbarProvider, { Snackbar } from './hooks/SnackbarProvider.jsx'
 import './index.css'
 
 ReactDOM.createRoot(document.getElementById('root')).render(
   <React.StrictMode>
-    <UserProvider>
-      <ThemeProvider>
-        <WeatherProvider>
-          <PlantProvider>
-            <RoomProvider>
-              <BrowserRouter basename={import.meta.env.VITE_BASE_PATH}>
-                <App />
-              </BrowserRouter>
-            </RoomProvider>
-          </PlantProvider>
-        </WeatherProvider>
-      </ThemeProvider>
-    </UserProvider>
+    <SnackbarProvider>
+      <UserProvider>
+        <ThemeProvider>
+          <WeatherProvider>
+            <PlantProvider>
+              <RoomProvider>
+                <BrowserRouter basename={import.meta.env.VITE_BASE_PATH}>
+                  <App />
+                </BrowserRouter>
+              </RoomProvider>
+            </PlantProvider>
+            <Snackbar />
+          </WeatherProvider>
+        </ThemeProvider>
+      </UserProvider>
+    </SnackbarProvider>
   </React.StrictMode>
 )
